refactor: lint 적용 및 패키지 분리
