+++ src/main/kotlin/blackjack/domain/card/CardNumber.kt
@@ -1,4 +1,6 @@
-package blackjack.domain
+package blackjack.domain.card
+
+import blackjack.domain.Score
 
 enum class CardNumber(
     val score: Score
