+++ src/main/kotlin/blackjack/domain/card/CardPattern.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 enum class CardPattern {
     HEART,
