+++ src/main/kotlin/blackjack/domain/card/CardShuffler.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 interface CardShuffler {
     fun shuffle(cards: List<Card>): List<Card>
