+++ src/main/kotlin/blackjack/domain/player/PlayerCards.kt
@@ -1,4 +1,6 @@
-package blackjack.domain
+package blackjack.domain.player
+
+import blackjack.domain.card.Card
 
 class PlayerCards {
