+++ src/main/kotlin/blackjack/domain/player/Players.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.player
 
 class Players(
     val values: List<Player>
