+++ src/main/kotlin/blackjack/view/CardViewCreator.kt
@@ -1,8 +1,8 @@
 package blackjack.view
 
-import blackjack.domain.Card
-import blackjack.domain.CardNumber
-import blackjack.domain.CardPattern
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardNumber
+import blackjack.domain.card.CardPattern
 
 object CardViewCreator {
     fun convert(card: Card): String {
@@ -14,7 +14,7 @@ object CardViewCreator {
     }
 
     private fun convertPattern(cardPattern: CardPattern): String {
-        return when(cardPattern) {
+        return when (cardPattern) {
             CardPattern.SPADE -> "스페이드"
             CardPattern.HEART -> "하트"
             CardPattern.DIAMOND -> "다이아몬드"
