+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,12 +1,11 @@
 package blackjack.view
 
-import blackjack.domain.Players
+import blackjack.domain.player.Players
 
 object OutputView {
 
     private const val DRAW_CARD_MESSAGE_FORMAT = "%s에게 %d장의 카드를 나누었습니다."
     private const val PLAYER_CARD_MESSAGE_FORMAT = "%s카드: %s"
-
 
     fun drawCardMessage(players: Players, initDrawCardCount: Int) {
         val playerNames = players.values.joinToString(", ") { it.name.value }
