+++ src/test/kotlin/blackjack/domain/player/PlayerCardsTest.kt
@@ -1,5 +1,6 @@
-package blackjack.domain
+package blackjack.domain.player
 
+import blackjack.domain.player.PlayerCards
 import blackjack.helper.DeckHelper
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
