+++ src/test/kotlin/blackjack/domain/player/PlayerNameTest.kt
@@ -1,5 +1,6 @@
-package blackjack.domain
+package blackjack.domain.player
 
+import blackjack.domain.player.PlayerName
 import io.kotest.assertions.throwables.shouldThrowWithMessage
 import io.kotest.core.spec.style.StringSpec
