+++ src/test/kotlin/blackjack/domain/player/PlayerTest.kt
@@ -1,5 +1,8 @@
-package blackjack.domain
+package blackjack.domain.player
 
+import blackjack.domain.player.Player
+import blackjack.domain.player.PlayerCards
+import blackjack.domain.player.PlayerName
 import blackjack.helper.DeckHelper
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
